+++ CommonDamageTypes/Revitalizing.cs
@@ -3,6 +3,7 @@ using R2API;
 using RoR2;
 using RoR2.Projectile;
 using System.Collections.Generic;
+using System.Linq;
 using UnityEngine;
 using UnityEngine.Networking;
 
@@ -41,6 +42,21 @@ namespace Skillsmas.DamageTypes
 			useDefaultValueConfigEntry: SkillsmasPlugin.ignoreBalanceConfig.bepinexConfigEntry
 		);
 
+		public static ConfigOptions.ConfigurableValue<float> energeticResonanceChance = ConfigOptions.ConfigurableValue.CreateFloat(
+			SkillsmasPlugin.PluginGUID,
+			SkillsmasPlugin.PluginName,
+			SkillsmasPlugin.config,
+			"Keyword: Revitalizing",
+			"Energetic Resonance Chance",
+			5f,
+			description: "Effect granted by the Energetic Resonance passive from the ArtificerExtended mod",
+			stringsToAffect: new List<string>
+			{
+				"KEYWORD_SKILLSMAS_ARTIFICEREXTENDED_ALTPASSIVE_WATER"
+			},
+			useDefaultValueConfigEntry: SkillsmasPlugin.ignoreBalanceConfig.bepinexConfigEntry
+		);
+
 		public override void OnPluginAwake()
         {
 			revitalizingDamageType = DamageAPI.ReserveDamageType();
@@ -67,6 +83,15 @@ namespace Skillsmas.DamageTypes
 			{
 				var healAmount = (fractionalHealing + attackerInfo.healthComponent.fullHealth * flatHealing / 100f) * damageInfo.procCoefficient;
 				var spawnEffect = Util.CheckRoll(100f * damageInfo.procCoefficient);
+				var cleanse = false;
+				if (SkillsmasPlugin.artificerExtendedEnabled)
+				{
+					if (SoftDependencies.ArtificerExtendedSupport.BodyHasAltPassive(attackerInfo.gameObject))
+					{
+						var waterPower = SoftDependencies.ArtificerExtendedSupport.GetWaterPower(attackerInfo.gameObject);
+						cleanse = Util.CheckRoll(energeticResonanceChance * waterPower, attackerInfo.master);
+					}
+				}
 				foreach (var teamMember in TeamComponent.GetTeamMembers(attackerInfo.teamIndex))
                 {
 					if (teamMember.body && teamMember.body.healthComponent && teamMember.body.healthComponent.alive)
@@ -82,6 +107,74 @@ namespace Skillsmas.DamageTypes
 							};
 							effectData.SetHurtBoxReference(teamMember.gameObject);
 							EffectManager.SpawnEffect(revitalizingEffectPrefab, effectData, true);
+						}
+
+						if (cleanse)
+						{
+							var debuffCount = 0;
+							var debuffsOnThisBody = new List<BuffIndex>();
+							var debuffDotIndices = new List<DotController.DotIndex>();
+							var nullDotIndex = (DotController.DotIndex)(-1);
+							foreach (var buffIndex in BuffCatalog.debuffBuffIndices)
+                            {
+								if (teamMember.body.HasBuff(buffIndex) && teamMember.body.timedBuffs.Any(x => x.buffIndex == buffIndex))
+								{
+									debuffCount++;
+									debuffsOnThisBody.Add(buffIndex);
+									debuffDotIndices.Add(nullDotIndex);
+								}
+                            }
+							var dotController = DotController.FindDotController(teamMember.gameObject);
+							if (dotController)
+                            {
+								for (var dotIndex = (DotController.DotIndex)0; dotIndex < (DotController.DotIndex)(DotAPI.VanillaDotCount + DotAPI.CustomDotCount); dotIndex++)
+								{
+									if (dotController.HasDotActive(dotIndex))
+									{
+										var dotDef = DotController.GetDotDef(dotIndex);
+										if (dotDef.associatedBuff != null)
+										{
+											debuffCount++;
+											debuffsOnThisBody.Add(dotDef.associatedBuff.buffIndex);
+											debuffDotIndices.Add(dotIndex);
+										}
+									}
+								}
+                            }
+							if (debuffCount > 0)
+                            {
+								var randomDebuff = RoR2Application.rng.RangeInt(0, debuffCount);
+								var randomDebuffIndex = debuffsOnThisBody[randomDebuff];
+								var randomDebuffDotIndex = debuffDotIndices[randomDebuff];
+								if (randomDebuffDotIndex != nullDotIndex)
+								{
+									if (dotController)
+									{
+										for (var dotStackIndex = 0; dotStackIndex < dotController.dotStackList.Count; dotStackIndex++)
+										{
+											var dotStack = dotController.dotStackList[dotStackIndex];
+											if (dotStack.dotIndex == randomDebuffDotIndex)
+											{
+												dotController.RemoveDotStackAtServer(dotStackIndex);
+												break;
+											}
+										}
+									}
+								}
+								else
+								{
+									for (var timedBuffIndex = 0; timedBuffIndex < teamMember.body.timedBuffs.Count; timedBuffIndex++)
+									{
+										var timedBuff = teamMember.body.timedBuffs[timedBuffIndex];
+										if (timedBuff.buffIndex == randomDebuffIndex)
+										{
+											teamMember.body.timedBuffs.RemoveAt(timedBuffIndex);
+											teamMember.body.RemoveBuff(randomDebuffIndex);
+											break;
+										}
+									}
+								}
+                            }
 						}
                     }
                 }